ModelAttribute: Add `ModelAttribute`s for a bunch of models&attributes
We want to define `ModelAttribute`s for an initial set of `Chouette`
models:

* Chouette::Route
* Chouette::JourneyPattern
* Chouette::VehicleJourney
* Chouette::Footnote
* Chouette::TimeTable
* Chouette::RoutingConstraintZone

I added these by looking at the `schema.rb` as well as some form views
for these models. Given that information, I made some guesses about what
attributes it might make sense to add here.

We want to end up with a collection of only editable attributes for
these models, so this excludes things like `created_at`, foreign keys,
and programmatically set fields.

This is my guess at what those fields are. We can refine the list later.

We'll be using this list for Compliance Control to get a set of fields
of a given type in order to create custom validations using those
fields.

Refs #4401