ModelAttribute: Add `.methods_by_class` method
Allows us to get all `ModelAttribute`s given a certain class name.

Needed an `#==` method to be able to test object equality more easily so
I added one in.

I don't like that the `klass` is a lowercase symbol and the `klass`
attribute to `.methods_by_class` is a constantized string. Want to
correct that and make them uniform. Only did it this way because that's
how it was suggested in the ticket.

Refs #4401