ModelAttribute spec: Move `#instance_variable_set` to `before` block
Recommendation from Robert. The `@__all__` is an implementation detail.
It's more fragile to use `#instance_variable_set` in each test. Instead,
put that part in a `before` block so that if we decide to change it, we
only have to do so in one place.

Now each test can blithely assume that `ModelAttribute.all` starts out
as an empty list.

Refs #4401