Workbench: Validate presence of `output`
Ensure all workbenches always have an associated `ReferentialSuite` by
validating presence of the `output` attribute.

In order to pass the 'should have a valid factory' test, we need to
define a `ReferentialSuite` factory and set up a default association on
`output` in the `Workbench` factory.

We also need to add set `output` when creating a `Workbench` in
`spec/support/referential.rb`, otherwise we get a validation error that
breaks the tests. Clean up the formatting of that call. In order to pass
a `ReferentialSuite`, I just create one above, but I'm not sure if I
should be using a `find_or_create` mechanism like the other associated
objects do above.

Refs #3520