Add ComplianceControlSetCopyWorker
A new worker that launches the work to copy a `ComplianceControlSet` to
a `ComplianceCheckSet`.

Currently only handles a single control set and referential pair.
Probably want to extend this to handle multiple lists of these objects.
Also need to add in a call to the Java API to validate the check.

Refs #4782