ComplianceControlSetCopyWorker spec: Always stub API call
In the first test, we weren't stubbing the Java API call, so it would
get executed for real. Put the HTTP stub in a `before(:each)` call so
that it gets stubbed for every test here.

Additionally, reorganise the tests a bit to clean them up. Move the HTTP
stub to a `let` to allow it to be called in a couple different places.
Since the HTTP stub depends on the `check_set` double, move that to a
`let` too. And have the `ComplianceControlSetCopier` mock always return
`check_set`, now that the app code depends on that return value.

Refs #4782