Migrate all foreign keys to `bigint` from `integer`
All our primary keys are supposed to be `bigint`s, and our foreign keys
should likewise be `bigint`s. But it's easy to forget this when creating
migrations, or some committers might not have been up to speed on the
convention. In any case, this changes all existing foreign keys to be
`bigint`s.

Need to get this out now so that Michel Etienne can use it to correctly
test the Java application for compliance check validations. That's why
I'm not including any future-looking checks or enforcements here. That
will be coming next.

Refs #4947