Add spec to ensure `schema.rb` keys are `bigint`s
We use `bigint`s by default for all primary and foreign keys in the
application.

Using `bigint`s provides two advantages:

* Matching the Java application, which expects `bigint`s
* Providing support for large integer identifiers

Here, we add a spec that checks the `schema.rb` file with line-by-line
string matching to validate that no primary or foreign keys use
non-`bigint` types.

The test is a bit rough at the moment, but it works against the
temporary migration included in this commit. It should be reworked,
though, to include line numbers of the problem fields.

Refs #4951