referentials_helper.rb#line_status: Fix style of text label
The label previously took on the colour of the icon (green or red) and
was set in Times/browser default.

It was set in the wrong style because the label was put inside the
`<span>` for the icon. To fix the problem, leave the icon `<span>` empty
and put the label just after it in a text DOM node. This makes the text
label render correctly, in black 14px Open Sans.

Also remove the unnecessary `cls` variables and ternaries as these are
already handled by the outer `if` condition.

Refs #5005