Referential spec: Add spec that two similar Referentials being persisted
Here we formalise a bug whereby two "identical" `Referential`s are able
to be saved even though Rails validation would normally prevent this. It
can happen when two `Referential`s are functionally identical (but have
different `slug`s, as this would trigger a Postgres error).

In practise, this happens when multiple imports of the same data are
launched at similar times.

We'll want to make this test pass by not allowing the identical
`Referential` to be created. This will be accomplished as agreed with
the team by a database table lock.

The threads and sleeps (in particular) are unfortunate necessities
needed to get the two `Referential`s to be saved at the same time.

TODO: Need to check this again with the threads & sleeps turned off to
confirm that this passes in that case (actually it doesn't), so this
test needs some modification in order to be correct.

Refs #5024