spec/support/referential: Rename `:js` filter to `:truncation`
This `before` hook allows tests to use the "truncation" database cleanup
strategy. It doesn't do anything explicitly JavaScript-related. Rather,
it was named JS because it was intended to be used for feature tests
with JavaScript. However, now that I need to use it for a couple tests
in `spec/models/referential_spec.rb`, the `:js` name doesn't make sense.

Rename the filter to something that says what it does, not how it should
be used.

Refs #5024