spec_helper.rb: Enable `truncation` test; Disable `wip` tests
I didn't notice this when I added the `truncation` filter to my
`referentials` table lock test, but the way the specs are configured, it
wouldn't be run by default with the rest of the tests.

Luc did notice, and suggested we change the filter options. We will now
always run tests with the `truncation: true` filter (removing the
`filter_run_excluding` config makes them enabled), as these include
JavaScript-dependent feature tests that we want to be sure to execute.

Additionally, we exclude tests with the `wip: true` filter, as it
doesn't really make sense to run these "work in progress" tests by
default.

Refs #5024