rails_helper.rb: Remove `use_transactional_fixtures = true`
We want `DatabaseCleaner` to handle our database cleanup because
sometimes we need to use a database truncation strategy instead of
transactions.

I need it to test the `Referential` table lock.

This line gets removed instead of the config being set to `false` here
because we defer to `spec_helper.rb` which already does that and does
the `DatabaseCleaner` stuff through `spec/support/referential.rb`.

Refs #5024