Referential: Lock table on :update
We had been locking the `referentials` table on :create, but we also
want to handle :update. Paired on this with Johan.

When I used:

    referential_2.metadatas << metadata_2

the referential was saved. To add the metadata without automatically
saving the referential + metadata, Johan suggested using the nested
attribute method:

    referential_2.metadatas_attributes = [metadata_2.attributes]

This allows us to add the metadata and still use the `#save` method to
lock the table.

Also change the callback from `before_validation` to `before_save`
because

    before_validation :lock_table, on: [:create, :update]

didn't work. That caused an error in our `expect`, as the `be_valid`
triggered the lock callback. To enable the callback on both :create and
:update, use a `before_save` instead.

Refs #5024