referential_lock_during_creation_spec: Skip truncation tests
Just discussed this with Alban. The truncation tests are breaking the
test suite, and I've spent way too long on this already. Take the specs
out of the suite for now until we can come up with a different way to
run these or test this behaviour without breaking everything.

The new task to fix this is #5295.

Refs #5024