Change `compliance_{controls,checks}` `control_attributes` to JSON
Make these columns JSON fields again. Do this because of bugs trying to
validate types with `hstore_accessor`
(https://github.com/devmynd/hstore_accessor/issues/78).

With a JSON field, types are preserved and we don't have to deal with
casting back and forth.

Previously we had converted to Hstore because the Java IEV application
used an older ORM version (maybe Hibernate) that didn't support the
Postgres JSON type. Because of the validation problems with Hstore, the
Java application has been updated to support JSON fields.

Thanks to this Stack Overflow post from 'a_horse_with_no_name' that
describes how to convert from Hstore to JSON types in Postgres:
https://stackoverflow.com/questions/33732529/postgres-how-to-convert-hstore-to-json-datatypes#comment55234342_33732529

Converting in this direction seems to me to be much safer because we're
not losing type information. All the values should be hashes of strings,
which should convert without issue into JSON.

Refs #5316