initializers/sidekiq: Disable retries on Sidekiq jobs
Alban recommended disabling retries on all Sidekiq jobs. In theory, none
of our jobs should be automatically retried (except maybe the
Reflex/Codifligne sync tasks).

Another way of doing this also exists
(https://stackoverflow.com/questions/28412913/disable-automatic-retry-with-activejob-used-with-sidekiq/28427385#28427385),
but this appeared to work, looking at the job's JSON parameters in the
Sidekiq console.

This came about because a timeout from Faraday caused a
`WorkbenchImport` job to fail and presumably get retried by Sidekiq. The
retry caused a second, duplicate `Referential` to be created.

Refs #5495