workbenches/_filters: Only show organisations in workgroup
We had been listing all `Organisation`s in the check-box list that
enables filtering `Referential`s by organisation.

There can be many organisations in the database, and users of one org
shouldn't necessarily see other organisations.

Only allow users to filter by those organisations that belong to their
`Workgroup`. To do that, We get the workgroup of the current workbench,
and build a list containing the organisations the workgroup.

Refs #5140