Imports#index: Move 'New import' button to action links
Begin the process of converting the `ImportDecorator` to the new action
links interface.

For now, only the "New import" button on the top right of the index page
is converted. The rest of the links are commented out for now and will
be treated in a subsequent commit.

Add a new `#class?` method to `AF83::Decorator::Link` that tells callers
whether a CSS class has been defined on the link. This allows us to
determine whether we should force a class or use the one provided in
`_page_header.html.slim`.

We need a way to specify the class of the link in the header because the
"New import" button is supposed to be blue (`.btn-primary`), but the
default for primary action links is white.

Refs #5586