AF83::Decorator::Link: Delete `#default_class`
This method was overriding classes that were added by `#add_class` in
addition to those added by `#class`. Obviously, the argument to
`#add_class` should be appended to the default.

That seemed like a pain to do, so instead I'm getting rid of the concept
of a default class. Instead, we're going to force all links to use the
template-defined styles for now. That goes in line with Luc's desire to
be more consistent in the links, as he stated yesterday.

Refs #5586