AF83::Decorator::Link: Add `data-method` to HTML options
Previously, we weren't correctly adding the HTTP method to the links in
the header. This is because it's called `link_method` in the class. Put
this value in `method`/`data-method` when converting the link to HTML so
that the links work properly.

Discovered this when deactivating and activating stop areas on
http://stif-boiv.dev:3000/stop_area_referentials/1/stop_areas/139429 .

Refs #5586