ImportResources#index: Get rid of `ImportResourceDecorator`
Seems like the object decorator for `ImportResource` isn't needed here.
It doesn't contain anything except for an empty `action_links` method,
which I'm guessing was to appease the table builder which used to expect
one.

Get rid of that decorator because it doesn't seem to be needed.

Also get rid of the `links` argument to the table builder because we
don't need it any more with the new action links interface.

Refs #5586