_select2.sass: Prevent drop-down arrow from overlapping selection
When the text of the current selection was very long, it could extend to
the end of the input, the farthest right:

    +-------------------+
    | Long text is long |
    +-------------------+

The problem is that the arrow indicating you can click on the input to
bring up a drop-down list of options would cover the "g" in "long" in
the above example (well, not exactly because of proportional width
characters, but you get the idea).

Fix the overlap by overriding Bootstrap Select2's style from
app/assets/stylesheets/vendor/select2-bootstrap.css:306-309, which looks
like this:

    .select2-container--bootstrap .select2-selection--single .select2-selection__rendered {
      color: #555555;
      padding: 0;
    }

Obviously, the padding is wrong. We need some right padding to give room
for the arrow. What are you thinking Bootstrap Select2?

Refs #5576