Revert "Prevent problem with StopArea#kind. Refs #5817"
This reverts commit c2bc391ee91cac70e726d188be97dd2323df0df2.

Thanks to 76abdbd66c16e6e1233685a5fa28a42cba5580d9, this temporary fix
can be reverted. We don't actually need `kind` selected here, it was
only added to appease the auto-initialisation of `kind` in `StopArea`.
Now that we don't set a default value for `kind` on initialisation, we
can remove it from this `select`.