Set all existing `StopArea`s with nil `kind`s to `:commercial`
The other migration for this in
`db/migrate/20180126134944_add_kind_to_stop_areas.rb` didn't handle
`kind IS NULL`. Since all the `kind`s were initialised to NULL when the
column was created, the `where` query didn't select any stop areas
because it was looking for those with some varchar value, and thus it
wasn't able to update any.

This ensures our existing records all have a `kind` in case we end up
using the field in a template at some point and our old data bugs out.

Refs #5817