ComplianceCheckSet: Add `.abort_old`
Copy over the `Import.abort_old` method to this class. For now I haven't
bothered with moving it to a consolidated module. We'll see, maybe
that's a better approach. For now I'm going with a relaxed copy-paste
solution until we decide that the increased complexity of the module is
worth the deduplication.

We need this method to be able to abort stale imports. We'll be doing
this, similarly to imports, in a regular cron job.

Refs #4758