TomTom: Use travelMode='bus'
For now, hard-code 'bus' as the travel mode. This feature is currently
only going to be used for Ouibus, so we want to be able to request in
'bus' mode, but for now don't have time to think about how to make it
more extensible to other travel modes and operators.

Refs #6095