TomTom#batch: Remove unwanted params
I had copy-pasted these from the example in the docs
(https://developer.tomtom.com/online-routing/online-routing-documentation/batch-routing),
but I don't think we want some of these. For `maxAlternatives`, it's
already 0 by default according to their docs, and the others don't
really make sense for us because we want a general routing instead of
one based on traffic or time of day.

Refs #6095