TomTom#batch: Populate WayCosts with distance & time from API
Look through the API response from `/batch` and extract the distance &
time values given to us by TomTom. These are inserted in the `WayCost`s
given to use in the argument to `#batch`.

At the end, we get a list of `WayCost`s that are completely filled with
distance & time values.

We need to split up this code a bit to make it more testable, but
already this seems to function correctly.

Refs #6095