Route: Don't run `#calculate_costs!` during a referential merge
This callback got triggered during a referential merge, but it isn't
needed at that point. It should only be executed as a result of a user
updating the route (via the web interface).

Using Alban's suggestion, we can detect that we're in the process of a
merge if the associated referential is in a `ReferentialSuite`. Use this
as a condition to run the callback.

Refs #6410