ReferentialVehicleJourneys: Use display name in line filter drop-down
Previously we had been showing the `name` field for the initial value of
an already selected line. However, that doesn't show all the information
that's normally given in the drop-down options. We instead have to use
the `#display_name` method to get the full display that people are used
to seeing in the drop-down.

This also means we can no longer `pluck`, as more fields are required to
assemble the display name. Instead, just do a simple `find` for the
record.

Refs #5889