Workbenches#show: Don't allow pending referentials to be selected
Disable the checkboxes next to pending referentials. This prevents them
from being deleted by the batch deletion operation.

In general it seems like we shouldn't be operating on pending
referentials, so if any new batch actions are added, it would make sense
to me for the checkboxes to stay disabled.

We want to disallow deletion of pending referentials because this can
put the data in an inconsistent state that requires manual intervention
to clean (I think).

Refs #6858