WayCost: Add `#calculate_distance`
This new method calculates the distance between the cost's departure and
arrival points.

It does this using RGeo's geographic factory using a spherical
calculation.

This allows us to get a bird's eye distance between the points in the
cost. The distance given to us by TomTom is locked to a road, which
means that point A on the left side of the highway and point B on the
right side can get a TomTom distance that's much further apart than the
actual bird's eye distance, because it includes offramp vehicle
circumnavigation to get to the other side of the highway.

We'll be using this distance value to compare against TomTom's. If
`#calculate_distance`'s result is under 500 meters, our custom logic (to
be implemented) will snap the distance to 1 kilometre.

Refs #6952